Rename CategoriesSources to Category, style resources into three responsive columns
